requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
tenacity>=8.2.0
imagehash>=4.3.1
numpy>=1.24.0
libsql-experimental
streamlit